Extract LP create_data_model/create_solver into conversion.py
