+++ python/cuopt_server/cuopt_server/utils/utils.py
@@ -1,15 +1,15 @@
-# SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2024-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
 import json
 import os
 
 from cuopt_server.utils.job_queue import SolverLPJob
-from cuopt_server.utils.linear_programming.data_definition import LPData
-from cuopt_server.utils.linear_programming.solver import (
+from cuopt_server.utils.linear_programming.conversion import (
     create_data_model as lp_create_data_model,
     create_solver as lp_create_solver,
 )
+from cuopt_server.utils.linear_programming.data_definition import LPData
 from cuopt_server.utils.routing.data_definition import OptimizedRoutingData
 from cuopt_server.utils.routing.solver import (
     create_data_model as routing_create_data_model,
